Migrate from Pyramid to Flask & Celery
We separate request handling to web and worker side and use Celery to
handle tasks from the queue.
