+++ Procfile
@@ -1,1 +1,2 @@
-web: gunicorn main:app
+web: gunicorn wsgi:app
+worker: celery -A tasks.celery worker -c 4
