+++ handlers.py
@@ -1,0 +1,6 @@
+from coordinators import MainCoordinator
+
+
+def populate_dispatcher(dispatcher):
+    main_coordinator = MainCoordinator()
+    dispatcher.add_handler(main_coordinator.conv_handler)
