+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "witty-bot"
+name = "bot"
 version = "0.1.0"
 description = ""
 authors = ["Sergey Karatach <karatach1998@yandex.ru>"]
@@ -12,7 +12,6 @@ wolframalpha = "^4.1"
 antlr4-python3-runtime = "^4.9"
 yadisk = "^1.2.14"
 PyYAML = "^5.4.1"
-pyramid = "^2.0"
 google-auth = "^1.27.1"
 google-auth-oauthlib = "^0.4.3"
 google-api-python-client = "^2.0.2"
@@ -23,6 +22,8 @@ PyPDF2 = "^1.26.0"
 lxml = "^4.6.2"
 mashumaro = "^2.0.2"
 gunicorn = "^20.1.0"
+Flask = "^1.1.2"
+celery = "^5.0.5"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.2"
