+++ viewcontrollers.py
@@ -1,6 +1,8 @@
+import io
 from itertools import chain
 from urllib.parse import urljoin
 
+import requests
 import yaml
 from telegram import (  # Update,
     InlineKeyboardButton,
@@ -66,11 +68,10 @@ class RussianRulesInlineDialogue(AbstractInlineDialogue):
 
         chat_id = update.effective_message.chat_id
         html_file_name = f"russian_rules-{chat_id}.html"
-        html_file_path = config.MEDIA_PATH / html_file_name
-        html_file_path.write_text(html)
         html_file_url = urljoin(
             config.APP_BASE_URL, f"/media/{html_file_name}"
         )
+        requests.post(html_file_url, files={'file': io.StringIO(html)})
         print(f'URL for {chapter_title}:', html_file_url)
 
         query.message.reply_text(
